+++ TwitterJSON/TJUser.swift
@@ -19,7 +19,7 @@ public class TJUser {
     /**
     The screen name of the user
     */
-    public var screenName: String!
+    public var screenName: String?
     
     /**
     User bio/description
@@ -51,6 +51,8 @@ public class TJUser {
         
         if let user = userInfo["name"].string {
             self.name = user
+        } else {
+            self.name = nil
         }
         
         if let screenName = userInfo["screen_name"].string {